Modify repository structure to specify Redshift-specific DAGs and SQL queries
Create a new folder under dags/ for Redshift example DAGs, and correspondingly change
SQL file names to refer to Redshift. This is to ready the repo for the addition of BigQuery
example DAGs, and it'll be easier for a new reader to understand the repo if it's partitioned
by the backend datastore that is being exhibited. Also, deleted an unused SQL file.