Get DAG working through minor bug fixes
Jinja templating was acting strange, possibly to do with f-strings,
so just put the SQL queries in directly as multi-line strings. Not the
prettiest but it works. Cleaned up some comments as well and added a
note about creating a GCP connection. Will go into more detail in the
README.