Add a simple Snowflake data quality pipeline
This pipeline is very similar to the Bigquery and Redshift EL 3 DAGs
in structure and usage. It serves as an example to show how data quality
can be added to a Snowflake DAG. It uses two operators, SnowflakeCheckOperator
and SnowflakeValueCheckOperator, that do not currently exist in the Snowflake provider;
a PR has been opened in Airflow to add these (link to the PR is in this PR's description).