Reconfigure row-level quality checks.
The double-for-loop and TaskGroup style of task creation did not conform
to best practices for DAG writing and was largely unnecssary. The main issue
was that it was loading a file during every DAG parse, which is ill-advised at
best. It was also unnecessary to do this, as the same work to load random rows
for quality checking could be done in the SQL query itself, offloading more compute
energy to the SQL server. So, that is what was done: the query was rewritten to
use a random pickup date to choose a row for quality checking, and only a single
TaskGroup now exists to hold 10 row quality checks. These quality checks must still
be in a TaskGroup, as the SQLCheckOperator only returns a single row, so multi-row
checking in the query is worthless. More discussion is available in PR comments.