Finish DAG writing and update dbt schema
Issues with BigQuery ID and dbt schema led to dropping the custom
test that was designed to fail, and designing the month test to
fail instead. The custom test for whatever reason was trying to
use the schema.project_id as the project ID and BigQuery did NOT
like that. Not sure what the fix is.

Additionally, the DAG is completed by adding a BigQuery copy table
operator. This copies a single table written by dbt test on failures
to a permament path. This is a not-so-great, but workable, solution
until AIP-42 gets merged and we can do some real dynamic task
generation based on XCOM return values. In preparation for that time,
BigQueryGetDatasetTablesOperator is left in commented-out.